refactor(Name, Names) : player 패키지로 이동
