+++ src/main/java/domain/player/Name.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.player;
 
 import java.util.Objects;
