+++ src/test/java/domain/card/PlayerCardsTest.java
@@ -1,6 +1,6 @@
 package domain.card;
 
-import domain.Name;
+import domain.player.Name;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
