+++ src/test/java/domain/game/RuleTest.java
@@ -1,10 +1,10 @@
 package domain.game;
 
-import domain.Name;
 import domain.card.Card;
 import domain.card.DealerCards;
 import domain.card.PlayerCards;
 import domain.card.Shape;
+import domain.player.Name;
 import domain.score.Status;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
