+++ src/test/java/domain/player/NameTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.player;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
