+++ src/test/java/domain/player/NamesTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.player;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
